export CSV optimization + UI matching figma
